requirements.txt
sanic==22.6.2
torch==1.13.1
diffusers==0.11.1
scipy
accelerate==0.12.0
triton
transformers
ftfy
bitsandbytes
natsort
torchvision
s3fs
python-dotenv